+++ example/main.go
@@ -1,57 +1,51 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
 	"github.com/beyond-blockchain/bbclib-go"
+	//"main/bbclib"
 	"io/ioutil"
+	"time"
 )
 
-func main() {
-	idConf := bbclib.BBcIdConfig{16, 12, 10, 8, 24}
-	bbclib.ConfigureIdLength(&idConf)
 
+func makeTransaction(num int) *bbclib.BBcTransaction {
 	assetGroupID := bbclib.GetIdentifierWithTimestamp("assetGroupID", 32)
 	u1 := bbclib.GetIdentifierWithTimestamp("user1", 32)
 	u2 := bbclib.GetIdentifierWithTimestamp("user2", 32)
 	keypair1, _ := bbclib.GenerateKeypair(bbclib.KeyTypeEcdsaP256v1, 4)
-	keypair2, _ := bbclib.GenerateKeypair(bbclib.KeyTypeEcdsaSECP256k1, 4)
+	keypair2, _ := bbclib.GenerateKeypair(bbclib.KeyTypeEcdsaP256v1, 4)
+	//keypair1 := bbclib.GenerateKeypair(bbclib.KeyTypeEcdsaP256v1, 4)
+	//keypair2 := bbclib.GenerateKeypair(bbclib.KeyTypeEcdsaP256v1, 4)
+
+	bodyString := fmt.Sprintf("teststring!!!!:%d", num)
+	filedat, _ := ioutil.ReadFile("./asset_test.go")
+	datobj := map[string]string{"param1": "aaa", "param2": "bbb", "param3": string(num)}
 
 	txobj := bbclib.MakeTransaction(3, 0, true)
-	bbclib.AddEventAssetBodyString(txobj, 0, &assetGroupID, &u1, "teststring!!!!!")
-	txobj.Events[0].AddMandatoryApprover(&u1)
-	filedat, _ := ioutil.ReadFile("./asset_test.go")
-	bbclib.AddEventAssetFile(txobj, 1, &assetGroupID, &u2, &filedat)
-	txobj.Events[1].AddMandatoryApprover(&u2)
-	datobj := map[string]string{"param1": "aaa", "param2": "bbb", "param3": "ccc"}
-	bbclib.AddEventAssetBodyObject(txobj, 2, &assetGroupID, &u1, &datobj)
-	txobj.Events[2].AddMandatoryApprover(&u1)
+	txobj.Events[0].SetAssetGroup(&assetGroupID).AddMandatoryApprover(&u1).CreateAsset(&u1, nil, bodyString)
+	txobj.Events[1].SetAssetGroup(&assetGroupID).AddMandatoryApprover(&u2).CreateAsset(&u2, &filedat, nil)
+	txobj.Events[2].SetAssetGroup(&assetGroupID).AddMandatoryApprover(&u1).CreateAsset(&u1, nil, &datobj)
+	txobj.AddWitness(&u1).AddWitness(&u2)
 
-	txobj.Witness.AddWitness(&u1)
-	txobj.Witness.AddWitness(&u2)
+	txobj.Sign(&u1, keypair1, false).Sign(&u2, keypair2, false)
+	return txobj
+}
 
-	bbclib.SignToTransaction(txobj, &u1, keypair1)
-	bbclib.SignToTransaction(txobj, &u2, keypair2)
 
-	fmt.Println("-------------transaction--------------")
-	fmt.Printf("%v", txobj.Stringer())
-	fmt.Println("--------------------------------------")
-
-	dat, err := txobj.Pack()
-	if err != nil {
-		fmt.Printf("failed to serialize transaction object (%v)", err)
+func main() {
+	//idConf := bbclib.BBcIdConfig{16, 12, 10, 8, 24}
+	//bbclib.ConfigureIdLength(&idConf)
+	start := time.Now().Unix()
+	for i := 0; i<10000; i++ {
+		txobj := makeTransaction(i)
+		if result := txobj.Signatures[0].Verify(txobj.Digest()); !result {
+			fmt.Println("Verification failed..")
+		}
+		if result := txobj.Signatures[1].Verify(txobj.Digest()); !result {
+			fmt.Println("Verification failed..")
+		}
 	}
-	fmt.Printf("Packed data: %x", dat)
-
-	obj2 := bbclib.BBcTransaction{}
-	obj2.Unpack(&dat)
-	obj2.Digest()
-	if result := obj2.Signatures[0].Verify(obj2.TransactionID); !result {
-		fmt.Println("Verification failed..")
-	}
-
-	if bytes.Compare(txobj.Events[0].Asset.AssetID, obj2.Events[0].Asset.AssetID) != 0 ||
-		bytes.Compare(txobj.TransactionID, obj2.TransactionID) != 0 {
-		fmt.Println("Not recovered correctly...")
-	}
+	end := time.Now().Unix()
+	fmt.Printf("Time: %vs\n", end-start)
 }