+++ transaction.go
@@ -129,41 +129,100 @@ func (p *BBcTransaction) SetIdLengthConf(conf * BBcIdConfig) {
 	p.TransactionIdLength = conf.TransactionIdLength
 }
 
-// AddEvent adds the BBcEvent object in the transaction object
-func (p *BBcTransaction) AddEvent(obj *BBcEvent) {
-	obj.SetIdLengthConf(&p.IdLengthConf)
-	p.Events = append(p.Events, obj)
-}
 
-// AddReference adds the BBcReference object in the transaction object
-func (p *BBcTransaction) AddReference(obj *BBcReference) {
+// AddEvent adds the BBcEvent object in the transaction object
+func (p *BBcTransaction) AddEvent(assetGroupId *[]byte, referenceIndices *[]int) *BBcTransaction {
+	obj := BBcEvent{Version: p.Version}
 	obj.SetIdLengthConf(&p.IdLengthConf)
-	p.References = append(p.References, obj)
-	obj.Transaction = p
+	if assetGroupId != nil {
+		obj.SetAssetGroup(assetGroupId)
+	}
+	if referenceIndices != nil {
+		for _, idx := range *referenceIndices {
+			obj.AddReferenceIndex(idx)
+		}
+	}
+	p.Events = append(p.Events, &obj)
+	return p
 }
 
 // AddRelation adds the BBcRelation object in the transaction object
-func (p *BBcTransaction) AddRelation(obj *BBcRelation) {
+func (p *BBcTransaction) AddRelation(assetGroupId *[]byte) *BBcTransaction {
+	obj := BBcRelation{Version: p.Version}
 	obj.SetIdLengthConf(&p.IdLengthConf)
-	obj.SetVersion(p.Version)
-	p.Relations = append(p.Relations, obj)
+	if assetGroupId != nil {
+		obj.SetAssetGroup(assetGroupId)
+	}
+	p.Relations = append(p.Relations, &obj)
+	return p
+}
+
+// AddReference adds the BBcReference object in the transaction object
+func (p *BBcTransaction) CreateReference(assetGroupID *[]byte, refTransaction *BBcTransaction, eventIdx int) *BBcTransaction {
+	obj := BBcReference{Version: p.Version, Transaction: p}
+	obj.SetIdLengthConf(&p.IdLengthConf)
+	obj.Add(assetGroupID, refTransaction, eventIdx)
+	p.References = append(p.References, &obj)
+	return p
+}
+
+
+// AddCrossRef sets the BBcCrossRef object in the transaction object
+func (p *BBcTransaction) CreateCrossRef(domainId, transactionId *[]byte) *BBcTransaction {
+	obj := BBcCrossRef{Version: p.Version}
+	obj.SetIdLengthConf(&p.IdLengthConf)
+	obj.Add(domainId, transactionId)
+	p.Crossref = &obj
+	return p
 }
 
 // AddWitness sets the BBcWitness object in the transaction object
-func (p *BBcTransaction) AddWitness(obj *BBcWitness) {
-	obj.SetIdLengthConf(&p.IdLengthConf)
-	p.Witness = obj
-	obj.Transaction = p
+func (p *BBcTransaction) AddWitness(userId *[]byte) *BBcTransaction {
+	if p.Witness == nil {
+		obj := BBcWitness{Version: p.Version}
+		obj.SetIdLengthConf(&p.IdLengthConf)
+		obj.Transaction = p
+		p.Witness = &obj
+	}
+	_ = p.Witness.AddWitness(userId)
+	return p
 }
 
-// AddCrossRef sets the BBcCrossRef object in the transaction object
-func (p *BBcTransaction) AddCrossRef(obj *BBcCrossRef) {
-	obj.SetIdLengthConf(&p.IdLengthConf)
-	p.Crossref = obj
+
+// AddSignature adds the BBcSignature object for the specified userID in the transaction object
+func (p *BBcTransaction) Sign(userId *[]byte, keyPair *KeyPair, noPubkey bool) *BBcTransaction {
+	obj := BBcSignature{Version: p.Version}
+	if noPubkey {
+		obj.SetPublicKeyInfo(uint32(keyPair.CurveType))
+	} else {
+		obj.SetPublicKeyByKeypair(keyPair)
+	}
+	signature, _ := p.doSign(keyPair)
+	obj.Signature = signature
+	obj.SignatureLen = uint32(len(signature)*8)
+
+	uid := make([]byte, int(p.IdLengthConf.UserIdLength))
+	copy(uid, *userId)
+	for i := range p.SigIndexedUsers {
+		if reflect.DeepEqual(p.SigIndexedUsers[i], uid) {
+			p.Signatures[i] = &obj
+			return p
+		}
+	}
+	if p.References != nil {
+		for i := range p.References {
+			if err := p.References[i].AddSignature(userId, &obj); err == nil {
+				return p
+			}
+		}
+	}
+	p.SigIndexedUsers = append(p.SigIndexedUsers, uid)
+	p.Signatures = append(p.Signatures, &obj)
+	return p
 }
 
 // AddSignature adds the BBcSignature object for the specified userID in the transaction object
-func (p *BBcTransaction) AddSignature(userID *[]byte, sig *BBcSignature) {
+func (p *BBcTransaction) AddSignatureObj(userID *[]byte, sig *BBcSignature) {
 	uid := make([]byte, int(p.IdLengthConf.UserIdLength))
 	copy(uid, *userID)
 	for i := range p.SigIndexedUsers {
@@ -175,6 +234,7 @@ func (p *BBcTransaction) AddSignature(userID *[]byte, sig *BBcSignature) {
 	p.SigIndexedUsers = append(p.SigIndexedUsers, uid)
 	p.Signatures = append(p.Signatures, sig)
 }
+
 
 // GetSigIndex reserves and returns the position (index) of the corespondent userID in the signature list
 func (p *BBcTransaction) GetSigIndex(userID []byte) int {
@@ -210,30 +270,13 @@ func (p *BBcTransaction) SetSigIndex(userID []byte, idx int) {
 }
 
 // Sign TransactionID using private key in the given keypair
-func (p *BBcTransaction) Sign(keypair *KeyPair) ([]byte, error) {
+func (p *BBcTransaction) doSign(keypair *KeyPair) ([]byte, error) {
 	digest := p.Digest()
 	signature := keypair.Sign(digest)
 	if signature == nil {
 		return nil, errors.New("fail to sign")
 	}
 	return signature, nil
-}
-
-// SignAndAdd signs to TransactionID and add BBcSignature in the transaction
-func (p *BBcTransaction) SignAndAdd(keypair *KeyPair, userID []byte, noPubkey bool) error {
-	digest := p.Digest()
-	signature := keypair.Sign(digest)
-	if signature == nil {
-		return errors.New("fail to sign")
-	}
-	idx := p.GetSigIndex(userID)
-	p.Signatures[idx].KeyType = uint32(keypair.CurveType)
-	p.Signatures[idx].SetSignature(&signature)
-	if ! noPubkey {
-		p.Signatures[idx].Pubkey = keypair.Pubkey
-		p.Signatures[idx].PubkeyLen = uint32(len(keypair.Pubkey) * 8)
-	}
-	return nil
 }
 
 // VerifyAll verifies TransactionID with all BBcSignature objects in the transaction